requirements.txt
# Core dependencies
PySide6>=6.5.0
redis>=4.5.0
requests>=2.28.0
python-dotenv>=1.0.0

# Busylight library
busylight-for-humans
# PyInstaller (for building)
pyinstaller>=5.12.0
pillow>=10.0.0

# Additional dependencies
altgraph==0.17.4
bitvector-for-humans==0.14.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
docopt==0.6.2
hidapi==0.14.0.post4
idna==3.10
loguru==0.7.3
macholib==1.16.3
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
pipreqs==0.4.13
Pygments==2.19.1
pyinstaller-hooks-contrib==2025.4
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
pyserial==3.5
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
rich==14.0.0
setuptools==80.7.1
shellingham==1.5.4
shiboken6==6.9.0
typer==0.15.4
typing_extensions==4.13.2
urllib3==2.4.0
webcolors==24.11.1
yarg==0.1.10

